move add mock data button to CatalogTable
