+++ app-config.yaml
@@ -50,6 +50,20 @@ newrelic:
     baseUrl: 'https://api.newrelic.com/v2'
     key: NEW_RELIC_REST_API_KEY
 
+catalog:
+  entities:
+    [
+      'artist-lookup-component.yaml',
+      'playback-order-component.yaml',
+      'podcast-api-component.yaml',
+      'queue-proxy-component.yaml',
+      'searcher-component.yaml',
+      'playback-lib-component.yaml',
+      'www-artist-component.yaml',
+      'shuffle-api-component.yaml',
+    ]
+  baseUrl: 'https://github.com/spotify/backstage/blob/master/packages/catalog-model/examples'
+
 auth:
   providers:
     google:
@@ -124,7 +138,7 @@ auth:
             env: AUTH_AUTH0_DOMAIN
     microsoft:
       development:
-        appOrigin: "http://localhost:3000/"
+        appOrigin: 'http://localhost:3000/'
         secure: false
         clientId:
           $secret:
